Failing tests for streaming encryption
